Move function into utils
Move reusable function `checkTagLineBreak` into utils
