+++ .vscode/markdownlint/utils.js
@@ -21,6 +21,54 @@ module.exports.addErrorAndInsertBlank = (onError, lineNumber, blankLineOffset=0)
 };
 
 /*
+ * Check a tag for a single line break between it and the enclosed content.
+ *
+ * @param {Object} tag The tag to check.
+ * @param {('opening'|'closing')} tagType Whether the tag is an opening or closing tag.
+ * @param {string} pattern A regex pattern that matches the tag
+ * @param {String[]} lines The lines in the document
+ * @param {addErrorCallback} onError The callback to add a markdownlint error and fix.
+ */
+module.exports.checkTagLineBreak = (tag, tagType, pattern, lines, onError) => {
+  if (tagType !== 'opening' && tagType !== 'closing') {
+    throw `The tag type for checkTagLineBreak must be either opening or closing: ${tagType}`;
+  }
+
+  // Check for lack of line break
+  const regex = (tagType === 'opening') ? `${pattern}$` : `^${pattern}`;
+  if (!tag.line.match(regex)) {
+    const match = tag.line.match(pattern);
+    const indexOfTag = match.index + 1;
+    const editColumn = (tagType === 'opening')
+      ? indexOfTag + match.at(0).length
+      : indexOfTag;
+    onError({
+      lineNumber: tag.lineNumber,
+      detail: 'No line break between content and tag',
+      fixInfo: {
+        insertText: '\n',
+        editColumn: editColumn,
+      },
+    });
+    return;
+  }
+
+  // Check for multiple line breaks
+  const lineNumberToCheck = (tagType === 'opening') ? tag.lineNumber : tag.lineNumber - 2;
+  if (this.isBlank(lines[lineNumberToCheck])) {
+    const lineNumberToFix = (tagType === 'opening') ? tag.lineNumber + 1 : tag.lineNumber - 1;
+    onError({
+      lineNumber: tag.lineNumber,
+      detail: 'Multiple line breaks between content and tag',
+      fixInfo: {
+        lineNumber: lineNumberToFix,
+        deleteCount: -1,
+      },
+    });
+  }
+};
+
+/*
  * Checks that the number of opening tags matches the number of closing tags. If
  * not, registers an error with markdownlint.
  *
